+++ tests/CcAcca.ProblemDetails.Helpers.Tests/EnsureSendAsync/ProblemDetailsResponse.cs
@@ -1,4 +1,6 @@
 using System.Net.Http;
+using System.Text.Json;
+using System.Threading;
 using System.Threading.Tasks;
 using CcAcca.ProblemDetails.Helpers.Tests.Fixtures;
 using Xunit;
@@ -14,11 +16,69 @@ namespace CcAcca.ProblemDetails.Helpers.Tests.EnsureSendAsync
         }
 
         [Fact]
-        public async void ResponseMessage()
+        public async void Options_CancellationToken_ResponseMessage()
         {
             // When
             var request = new HttpRequestMessage(HttpMethod.Get, Url);
-            async Task Act() => await Client.EnsureSendAsync(request);
+            async Task Act() => await Client
+                .EnsureSendAsync(request, HttpCompletionOption.ResponseHeadersRead, CancellationToken.None);
+
+            // then
+            await AssertThrows(Act, Problem);
+        }
+
+        [Fact]
+        public async void CancellationToken_ResponseMessage()
+        {
+            // When
+            var request = new HttpRequestMessage(HttpMethod.Get, Url);
+            async Task Act() => await Client.EnsureSendAsync(request, CancellationToken.None);
+
+            // then
+            await AssertThrows(Act, Problem);
+        }
+
+        [Fact]
+        public async void Options_CancellationToken_Return_Type()
+        {
+            // When
+            var request = new HttpRequestMessage(HttpMethod.Get, Url);
+            async Task Act() => await Client
+                .EnsureSendAsync<ExampleModel>(request, new JsonSerializerOptions(), CancellationToken.None);
+
+            // then
+            await AssertThrows(Act, Problem);
+        }
+
+        [Fact]
+        public async void CancellationToken_Return_Type()
+        {
+            // When
+            var request = new HttpRequestMessage(HttpMethod.Get, Url);
+            async Task Act() => await Client.EnsureSendAsync<ExampleModel>(request, CancellationToken.None);
+
+            // then
+            await AssertThrows(Act, Problem);
+        }
+
+        [Fact]
+        public async void Options_CancellationToken_Return_Object()
+        {
+            // When
+            var request = new HttpRequestMessage(HttpMethod.Get, Url);
+            async Task Act() => await Client
+                .EnsureSendAsync(request, typeof(ExampleModel), new JsonSerializerOptions(), CancellationToken.None);
+
+            // then
+            await AssertThrows(Act, Problem);
+        }
+
+        [Fact]
+        public async void CancellationToken_Return_Object()
+        {
+            // When
+            var request = new HttpRequestMessage(HttpMethod.Get, Url);
+            async Task Act() => await Client.EnsureSendAsync(request, typeof(ExampleModel), CancellationToken.None);
 
             // then
             await AssertThrows(Act, Problem);