+++ tests/CcAcca.ProblemDetails.Helpers.Tests/EnsureSendAsync/SuccessResponse.cs
@@ -36,5 +36,107 @@ namespace CcAcca.ProblemDetails.Helpers.Tests.EnsureSendAsync
             // then
             response.StatusCode.Should().Be(HttpStatusCode.OK);
         }
+
+        [Fact]
+        public async void CancellationToken_ResponseMessage()
+        {
+            // when
+            var request = new HttpRequestMessage(HttpMethod.Get, Url);
+            var response = await Client.EnsureSendAsync(request, CancellationToken.None);
+
+            // then
+            response.StatusCode.Should().Be(HttpStatusCode.OK);
+        }
+
+        [Fact]
+        public async void Result_Type()
+        {
+            // when
+            var request = new HttpRequestMessage(HttpMethod.Get, Url);
+            var result = await Client.EnsureSendAsync<ExampleModel>(request);
+
+            // then
+            result.Should().BeEquivalentTo(Model);
+        }
+
+        [Fact]
+        public async void JsonSerializerOptions_Result_Type()
+        {
+            // when
+            var request = new HttpRequestMessage(HttpMethod.Get, Url);
+            var result = await Client.EnsureSendAsync<ExampleModel>(request, Options);
+
+            // then
+            result.Should().BeEquivalentTo(Model);
+        }
+
+        [Fact]
+        public async void JsonSerializerOptions_CancellationToken_Result_Type()
+        {
+            // when
+            var request = new HttpRequestMessage(HttpMethod.Get, Url);
+            var result =
+                await Client.EnsureSendAsync<ExampleModel>(request, Options, CancellationToken.None);
+
+            // then
+            result.Should().BeEquivalentTo(Model);
+        }
+
+        [Fact]
+        public async void CancellationToken_Result_Type()
+        {
+            // when
+            var request = new HttpRequestMessage(HttpMethod.Get, Url);
+            var result =
+                await Client.EnsureSendAsync<ExampleModel>(request, CancellationToken.None);
+
+            // then
+            result.Should().BeEquivalentTo(Model);
+        }
+
+        [Fact]
+        public async void Result_Object()
+        {
+            // when
+            var request = new HttpRequestMessage(HttpMethod.Get, Url);
+            var result = await Client.EnsureSendAsync(request, typeof(ExampleModel));
+
+            // then
+            result.Should().BeEquivalentTo(Model);
+        }
+
+        [Fact]
+        public async void JsonSerializerOptions_Result_Object()
+        {
+            // when
+            var request = new HttpRequestMessage(HttpMethod.Get, Url);
+            var result = await Client.EnsureSendAsync(request, typeof(ExampleModel), Options);
+
+            // then
+            result.Should().BeEquivalentTo(Model);
+        }
+
+        [Fact]
+        public async void JsonSerializerOptions_CancellationToken_Result_Object()
+        {
+            // when
+            var request = new HttpRequestMessage(HttpMethod.Get, Url);
+            var result =
+                await Client.EnsureSendAsync(request, typeof(ExampleModel), Options, CancellationToken.None);
+
+            // then
+            result.Should().BeEquivalentTo(Model);
+        }
+
+        [Fact]
+        public async void CancellationToken_Result_Object()
+        {
+            // when
+            var request = new HttpRequestMessage(HttpMethod.Get, Url);
+            var result = await Client.EnsureSendAsync(request, typeof(ExampleModel), CancellationToken.None);
+
+            // then
+            result.Should().BeEquivalentTo(Model);
+        }
     }
 }